feat(kubernetes): add cloudnative-pg resources for database management
Add new Kubernetes manifests for managing the cloudnative-pg application within the database namespace. This includes the creation of an ExternalSecret, a Subscription for the cloudnative-pg operator, and the necessary Kustomization files to orchestrate the deployment and configuration. The ExternalSecret is configured to use a ClusterSecretStore with one password connect, and the Subscription is set up to automatically approve install plans from the certified-operators source in the openshift-marketplace. Additionally, a new Namespace manifest is added with pruning disabled. These changes are part of the ongoing efforts to automate and manage database resources effectively.